Do not squash handlers and signals chainmap before sending it to the external signaller
Squashing breaks handlers sorting, so it has been removed. Squashing helped
the external signaller to remove an handler not defined in the top level
class, because it's impossible to do that in a ChainMap.

The way now to achieve the same result (_check_local_handlers() is called just
after ExternalSignaller.register_class and it will check if a just defined
handler refers to an existent signal) it to add a signal with the name of the
signal name referenced by the handler and with value of None